requirements.txt
uvicorn
requests
fastapi
httpx
fastmcp
openai